feat: merged wallet providers into one entrypoint
