+++ typescript/agentkit/src/network/svm.ts
@@ -8,6 +8,11 @@ export type SOLANA_NETWORK_ID =
   | typeof SOLANA_MAINNET_NETWORK_ID
   | typeof SOLANA_TESTNET_NETWORK_ID
   | typeof SOLANA_DEVNET_NETWORK_ID;
+export const SOLANA_NETWORK_IDS: SOLANA_NETWORK_ID[] = [
+  SOLANA_MAINNET_NETWORK_ID,
+  SOLANA_TESTNET_NETWORK_ID,
+  SOLANA_DEVNET_NETWORK_ID,
+];
 
 // AgentKit Protocol Family
 export const SOLANA_PROTOCOL_FAMILY = "svm";
