+++ typescript/agentkit/src/wallet-providers/index.ts
@@ -1,8 +1,10 @@
 export * from "./walletProvider";
 export * from "./evmWalletProvider";
 export * from "./viemWalletProvider";
-export * from "./cdpV2WalletProvider";
+export * from "./cdpV2EvmWalletProvider";
 export * from "./cdpV2SolanaWalletProvider";
+export * from "./cdpV2WalletProvider";
+export * from "./cdpV2Shared";
 export * from "./cdpWalletProvider";
 export * from "./svmWalletProvider";
 export * from "./smartWalletProvider";
