+++ typescript/examples/langchain-cdp-v2-chatbot/.prettierignore
