--- typescript/examples/langchain-cdp-v2-solana-chatbot/.env-local
@@ -1,11 +1,0 @@
-# Required
-OPENAI_API_KEY=
-CDP_API_KEY_ID=
-CDP_API_KEY_SECRET=
-CDP_WALLET_SECRET=
-
-# Optional
-## Used for seeding account creation
-IDEMPOTENCY_KEY=
-## Used for recreationg existing accounts
-ADDRESS=
