--- typescript/examples/langchain-cdp-v2-solana-chatbot/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "parser": "@typescript-eslint/parser",
-  "extends": ["../../.eslintrc.base.json"]
-}
