--- typescript/examples/langchain-cdp-v2-solana-chatbot/.prettierignore
@@ -1,7 +1,0 @@
-docs/
-dist/
-coverage/
-.github/
-src/client
-**/**/*.json
-*.md
