--- typescript/examples/langchain-cdp-v2-solana-chatbot/.prettierrc
@@ -1,11 +1,0 @@
-{
-  "tabWidth": 2,
-  "useTabs": false,
-  "semi": true,
-  "singleQuote": false,
-  "trailingComma": "all",
-  "bracketSpacing": true,
-  "arrowParens": "avoid",
-  "printWidth": 100,
-  "proseWrap": "never"
-}
