--- typescript/examples/langchain-cdp-v2-solana-chatbot/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "preserveSymlinks": true,
-    "outDir": "./dist",
-    "rootDir": "."
-  },
-  "include": ["*.ts"]
-}
